pyproject.toml
[project]
name = "f1-manager-strategy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "tabulate>=0.9.0",
    "tqdm>=4.66.0",
]
